fix some some
